fix: resolve all clippy lints across workspace
- Use #[default] derive instead of manual Default impls
- Remove needless borrows in sqlx .bind() calls
- Use .first() instead of .get(0), .next_back() instead of .last()
- Collapse nested if statements and merge identical branches
- Remove unnecessary .into() conversions and casts
- Use .is_some() instead of redundant pattern matching
- Initialize structs with field values instead of field reassignment

+++ services/signapps-ai/src/qdrant/client.rs
@@ -238,7 +238,7 @@ impl QdrantService {
         let result = info.result.ok_or_else(|| Error::Internal("No collection info".to_string()))?;
 
         // Get points count from segments_count as approximation
-        let segments_count = result.segments_count as u64;
+        let segments_count = result.segments_count;
         let status = match result.status() {
             qdrant_client::qdrant::CollectionStatus::Green => "green",
             qdrant_client::qdrant::CollectionStatus::Yellow => "yellow",
